docs(PasswordInput): migrate to Storybook
