+++ src/components/inputs/passwordInput/PasswordInput.mdx
@@ -1,0 +1,15 @@
+import { Meta, Controls, Stories, Title, Primary } from '@storybook/addon-docs/blocks';
+import * as PasswordInputStories from './PasswordInput.stories';
+
+<Meta of={PasswordInputStories} />
+
+<Title />
+
+<Primary />
+
+<Controls />
+
+See [Input](#/Components/Inputs/Input) for details about PasswordInput component 
+
+<Stories />
+
